requirements.txt
tensorflow==1.14
scipy
scikit-learn
opencv-python
h5py
matplotlib
Pillow
requests
psutil
imageio